+++ pkg/sql/colexec/colexecproj/projection_ops_test.go
@@ -102,70 +102,6 @@ func TestProjDivFloat64Float64Op(t *testing.T) {
 		})
 }
 
-func TestGetProjectionConstOperator(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	defer log.Scope(t).Close(t)
-	binOp := treebin.MakeBinaryOperator(treebin.Mult)
-	var input colexecop.Operator
-	colIdx := 3
-	inputTypes := make([]*types.T, colIdx+1)
-	inputTypes[colIdx] = types.Float
-	constVal := 31.37
-	constArg := tree.NewDFloat(tree.DFloat(constVal))
-	outputIdx := 5
-	op, err := GetProjectionRConstOperator(
-		testAllocator, inputTypes, types.Float, types.Float, binOp, input, colIdx,
-		constArg, outputIdx, nil /* EvalCtx */, nil /* BinFn */, nil, /* cmpExpr */
-	)
-	if err != nil {
-		t.Error(err)
-	}
-	expected := &projMultFloat64Float64ConstOp{
-		projConstOpBase: projConstOpBase{
-			OneInputHelper: colexecop.MakeOneInputHelper(op.(*projMultFloat64Float64ConstOp).Input),
-			allocator:      testAllocator,
-			colIdx:         colIdx,
-			outputIdx:      outputIdx,
-		},
-		constArg: constVal,
-	}
-	if !reflect.DeepEqual(op, expected) {
-		t.Errorf("got %+v,\nexpected %+v", op, expected)
-	}
-}
-
-func TestGetProjectionConstMixedTypeOperator(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	defer log.Scope(t).Close(t)
-	cmpOp := treecmp.MakeComparisonOperator(treecmp.GE)
-	var input colexecop.Operator
-	colIdx := 3
-	inputTypes := make([]*types.T, colIdx+1)
-	inputTypes[colIdx] = types.Int
-	constVal := int16(31)
-	constArg := tree.NewDInt(tree.DInt(constVal))
-	outputIdx := 5
-	op, err := GetProjectionRConstOperator(
-		testAllocator, inputTypes, types.Int2, types.Int, cmpOp, input, colIdx,
-		constArg, outputIdx, nil /* EvalCtx */, nil /* BinFn */, nil, /* cmpExpr */
-	)
-	if err != nil {
-		t.Error(err)
-	}
-	expected := &projGEInt64Int16ConstOp{
-		projConstOpBase: projConstOpBase{
-			OneInputHelper: colexecop.MakeOneInputHelper(op.(*projGEInt64Int16ConstOp).Input),
-			allocator:      testAllocator,
-			colIdx:         colIdx,
-			outputIdx:      outputIdx,
-		},
-		constArg: constVal,
-	}
-	if !reflect.DeepEqual(op, expected) {
-		t.Errorf("got %+v,\nexpected %+v", op, expected)
-	}
-}
-
 // TestRandomComparisons runs comparisons against all scalar types with random
 // non-null data verifying that the result of Datum.Compare matches the result
 // of the exec projection.
