+++ pkg/sql/colexec/colexecprojconst/BUILD.bazel
@@ -1,13 +1,17 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 load("//pkg/sql/colexecop:EXECGEN.bzl", "eg_go_filegroup", "gen_eg_go_rules")
+load("//pkg/sql/colexec:COLEXEC.bzl", "gen_default_cmp_proj_const_rule")
 load("//pkg/testutils/buildutil:buildutil.bzl", "disallowed_imports_test")
 
 go_library(
-    name = "colexecproj",
+    name = "colexecprojconst",
     srcs = [
+        "like_ops.go",
+        "proj_const_ops_base.go",
+        ":gen-default-cmp-proj-const-op",  # keep
         ":gen-exec",  # keep
     ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/sql/colexec/colexecproj",  # keep
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/colexec/colexecprojconst",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/col/coldata",  # keep
@@ -16,71 +20,45 @@ go_library(
         "//pkg/server/telemetry",  # keep
         "//pkg/sql/colconv",  # keep
         "//pkg/sql/colexec/colexecbase",  # keep
-        "//pkg/sql/colexec/colexeccmp",  # keep
-        "//pkg/sql/colexec/colexecutils",  # keep
+        "//pkg/sql/colexec/colexeccmp",
+        "//pkg/sql/colexec/colexecutils",
         "//pkg/sql/colexec/execgen",  # keep
         "//pkg/sql/colexecerror",  # keep
-        "//pkg/sql/colexecop",  # keep
-        "//pkg/sql/colmem",  # keep
+        "//pkg/sql/colexecop",
+        "//pkg/sql/colmem",
         "//pkg/sql/execinfra",  # keep
-        "//pkg/sql/sem/tree",  # keep
+        "//pkg/sql/sem/tree",
         "//pkg/sql/sem/tree/treebin",  # keep
         "//pkg/sql/sem/tree/treecmp",  # keep
         "//pkg/sql/sqltelemetry",  # keep
-        "//pkg/sql/types",  # keep
+        "//pkg/sql/types",
         "//pkg/util/duration",  # keep
         "//pkg/util/json",  # keep
         "@com_github_cockroachdb_apd_v3//:apd",  # keep
-        "@com_github_cockroachdb_errors//:errors",  # keep
+        "@com_github_cockroachdb_errors//:errors",
     ],
 )
 
 go_test(
-    name = "colexecproj_test",
-    srcs = [
-        "default_cmp_op_test.go",
-        "inject_setup_test.go",
-        "main_test.go",
-        "projection_ops_test.go",
-    ],
-    embed = [":colexecproj"],  # keep
-    tags = ["no-remote"],
+    name = "colexecprojconst_test",
+    srcs = ["projection_ops_test.go"],
+    embed = [":colexecprojconst"],
     deps = [
-        "//pkg/col/coldata",
-        "//pkg/col/coldataext",
-        "//pkg/col/coldatatestutils",
-        "//pkg/settings/cluster",
-        "//pkg/sql/colconv",
-        "//pkg/sql/colexec/colbuilder",
-        "//pkg/sql/colexec/colexecargs",
-        "//pkg/sql/colexec/colexectestutils",
-        "//pkg/sql/colexec/execgen",
-        "//pkg/sql/colexecerror",
         "//pkg/sql/colexecop",
-        "//pkg/sql/colmem",
-        "//pkg/sql/execinfra",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sem/tree/treebin",
         "//pkg/sql/sem/tree/treecmp",
         "//pkg/sql/types",
-        "//pkg/testutils/skip",
         "//pkg/util/leaktest",
         "//pkg/util/log",
-        "//pkg/util/mon",
-        "//pkg/util/randutil",
-        "@com_github_stretchr_testify//assert",
-        "@com_github_stretchr_testify//require",
     ],
 )
 
-# Export the template because it is used by another target in colexecprojconst
-# package.
-exports_files(["default_cmp_proj_ops_tmpl.go"])
-
 # Map between target name and relevant template.
 targets = [
-    ("default_cmp_proj_op.eg.go", "default_cmp_proj_ops_tmpl.go"),
-    ("proj_non_const_ops.eg.go", "proj_non_const_ops_tmpl.go"),
+    ("proj_const_left_ops.eg.go", "proj_const_ops_tmpl.go"),
+    ("proj_const_right_ops.eg.go", "proj_const_ops_tmpl.go"),
+    ("proj_like_ops.eg.go", "proj_const_ops_tmpl.go"),
 ]
 
 # Define a file group for all the .eg.go targets.
@@ -92,14 +70,20 @@ eg_go_filegroup(
 # Define gen rules for individual eg.go files.
 gen_eg_go_rules(targets)
 
+# Special-case the gen rule for default_cmp_proj_const_op.eg.go.
+gen_default_cmp_proj_const_rule(
+    name = "gen-default-cmp-proj-const-op",
+    target = "default_cmp_proj_const_op.eg.go",
+)
+
 disallowed_imports_test(
-    "colexecproj",
+    "colexecprojconst",
     [
         "//pkg/sql/colexec",
         "//pkg/sql/colexec/colexecagg",
         "//pkg/sql/colexec/colexechash",
         "//pkg/sql/colexec/colexecjoin",
-        "//pkg/sql/colexec/colexecprojconst",
+        "//pkg/sql/colexec/colexecproj",
         "//pkg/sql/colexec/colexecsel",
         "//pkg/sql/colexec/colexecwindow",
     ],
