+++ pkg/sql/colexec/colexecprojconst/proj_const_right_ops.eg.go
@@ -7,7 +7,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package colexecproj
+package colexecprojconst
 
 import (
 	"bytes"
