+++ pkg/sql/colexec/colexecprojconst/proj_like_ops.eg.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package colexecproj
+package colexecprojconst
 
 import (
 	"bytes"
