+++ pkg/sql/colexec/colexecspan/BUILD.bazel
@@ -94,6 +94,7 @@ disallowed_imports_test(
         "//pkg/sql/colexec/colexechash",
         "//pkg/sql/colexec/colexecjoin",
         "//pkg/sql/colexec/colexecproj",
+        "//pkg/sql/colexec/colexecprojconst",
         "//pkg/sql/colexec/colexecsel",
         "//pkg/sql/colexec/colexecwindow",
     ],
