+++ pkg/sql/colexec/execgen/cmd/execgen/default_cmp_proj_ops_gen.go
@@ -77,7 +77,7 @@ func init() {
 	registerGenerator(
 		getDefaultCmpProjOps(
 			defaultCmpProjOpOverload{
-				TargetPkg:    "colexecproj",
+				TargetPkg:    "colexecprojconst",
 				IsRightConst: true,
 				Kind:         "RConst",
 			}),
