+++ pkg/sql/colexec/execgen/cmd/execgen/like_ops_gen.go
@@ -24,8 +24,10 @@ import (
 // are several different implementations which may be chosen depending on the
 // complexity of the LIKE pattern.
 //
-// likeTemplate needs to be used as a format string expecting exactly one %s
-// arguments that describes the type of the operator (either "sel" or "proj").
+// likeTemplate needs to be used as a format string expecting exactly two %s
+// arguments:
+// [1]: is the package suffix (either "sel" or "projconst")
+// [2]: is the type of the operator (either "sel" or "proj").
 const likeTemplate = `
 // Copyright 2019 The Cockroach Authors.
 //
@@ -48,12 +50,15 @@ import (
 )
 
 {{range .}}
-{{template "%[1]sConstOp" .}}
+{{template "%[2]sConstOp" .}}
 {{end}}
 `
 
 func genLikeOps(
 	tmplGetter func(inputFileContents string) (*template.Template, error),
+	// pkgSuffix is the suffix of the package for the operator to be generated
+	// in (either "sel" or "projconst").
+	pkgSuffix string,
 	// opType is the type of the operator to be generated (either "sel" or
 	// "proj").
 	opType string,
@@ -63,7 +68,7 @@ func genLikeOps(
 		if err != nil {
 			return err
 		}
-		tmpl, err = tmpl.Parse(fmt.Sprintf(likeTemplate, opType))
+		tmpl, err = tmpl.Parse(fmt.Sprintf(likeTemplate, pkgSuffix, opType))
 		if err != nil {
 			return err
 		}
@@ -147,6 +152,6 @@ func init() {
 		projTemplate := replaceProjConstTmplVariables(inputFileContents, false /* isConstLeft */)
 		return template.New("proj_like_ops").Funcs(template.FuncMap{"buildDict": buildDict}).Parse(projTemplate)
 	}
-	registerGenerator(genLikeOps(getProjectionOpsTmpl, "proj"), "proj_like_ops.eg.go", projConstOpsTmpl)
-	registerGenerator(genLikeOps(getSelectionOpsTmpl, "sel"), "sel_like_ops.eg.go", selectionOpsTmpl)
+	registerGenerator(genLikeOps(getProjectionOpsTmpl, "projconst", "proj"), "proj_like_ops.eg.go", projConstOpsTmpl)
+	registerGenerator(genLikeOps(getSelectionOpsTmpl, "sel", "sel"), "sel_like_ops.eg.go", selectionOpsTmpl)
 }
