+++ pkg/sql/opt/exec/execbuilder/testdata/vectorize_local
@@ -281,7 +281,7 @@ EXPLAIN (VEC) SELECT stddev((t46404_1.c0 > ANY (0, 0))::INT) FROM t46404_0, t464
 └ Node 1
   └ *colexec.hashAggregator
     └ *colexecbase.castBoolIntOp
-      └ *colexecproj.defaultCmpRConstProjOp
+      └ *colexecprojconst.defaultCmpRConstProjOp
         └ *colexecjoin.crossJoiner
           ├ *colfetcher.ColBatchScan
           └ *colfetcher.ColBatchScan
@@ -328,7 +328,7 @@ EXPLAIN (VEC) SELECT _int2 * _int2 FROM ints WHERE _int4 + _int4 = _int8 + 2
     └ *colexecbase.castInt2IntOp
       └ *colexecbase.castInt2IntOp
         └ *colexecsel.selEQInt64Int64Op
-          └ *colexecproj.projPlusInt64Int64ConstOp
+          └ *colexecprojconst.projPlusInt64Int64ConstOp
             └ *colexecproj.projPlusInt64Int64Op
               └ *colexecbase.castInt4IntOp
                 └ *colexecbase.castInt4IntOp
