+++ pkg/sql/opt/exec/execbuilder/testdata/vectorize_overloads
@@ -169,7 +169,7 @@ EXPLAIN (VEC) SELECT _inet - 1 FROM many_types
 ----
 │
 └ Node 1
-  └ *colexecproj.projMinusDatumInt64ConstOp
+  └ *colexecprojconst.projMinusDatumInt64ConstOp
     └ *colfetcher.ColBatchScan
 
 query T
@@ -186,7 +186,7 @@ EXPLAIN (VEC) SELECT 2 + _inet FROM many_types
 ----
 │
 └ Node 1
-  └ *colexecproj.projPlusDatumInt64ConstOp
+  └ *colexecprojconst.projPlusDatumInt64ConstOp
     └ *colfetcher.ColBatchScan
 
 query T
@@ -242,7 +242,7 @@ EXPLAIN (VEC) SELECT _int << 1 FROM many_types
 ----
 │
 └ Node 1
-  └ *colexecproj.projLShiftInt64Int64ConstOp
+  └ *colexecprojconst.projLShiftInt64Int64ConstOp
     └ *colfetcher.ColBatchScan
 
 query T
@@ -250,7 +250,7 @@ EXPLAIN (VEC) SELECT _int >> 1 FROM many_types
 ----
 │
 └ Node 1
-  └ *colexecproj.projRShiftInt64Int64ConstOp
+  └ *colexecprojconst.projRShiftInt64Int64ConstOp
     └ *colfetcher.ColBatchScan
 
 query T
@@ -258,7 +258,7 @@ EXPLAIN (VEC) SELECT _varbit << 4 FROM many_types
 ----
 │
 └ Node 1
-  └ *colexecproj.projLShiftDatumInt64ConstOp
+  └ *colexecprojconst.projLShiftDatumInt64ConstOp
     └ *colfetcher.ColBatchScan
 
 query T
@@ -293,7 +293,7 @@ EXPLAIN (VEC) SELECT _varbit >> 4 FROM many_types
 ----
 │
 └ Node 1
-  └ *colexecproj.projRShiftDatumInt64ConstOp
+  └ *colexecprojconst.projRShiftDatumInt64ConstOp
     └ *colfetcher.ColBatchScan
 
 query T
@@ -353,7 +353,7 @@ EXPLAIN (VEC) SELECT _json -> 2 FROM many_types
 ----
 │
 └ Node 1
-  └ *colexecproj.projJSONFetchValJSONInt64ConstOp
+  └ *colexecprojconst.projJSONFetchValJSONInt64ConstOp
     └ *colfetcher.ColBatchScan
 
 query T
@@ -361,8 +361,8 @@ EXPLAIN (VEC) SELECT _json -> 2 -> 'a' FROM many_types
 ----
 │
 └ Node 1
-  └ *colexecproj.projJSONFetchValJSONBytesConstOp
-    └ *colexecproj.projJSONFetchValJSONInt64ConstOp
+  └ *colexecprojconst.projJSONFetchValJSONBytesConstOp
+    └ *colexecprojconst.projJSONFetchValJSONInt64ConstOp
       └ *colfetcher.ColBatchScan
 
 query T
@@ -370,7 +370,7 @@ EXPLAIN (VEC) SELECT _json #> _stringarray, _json #> '{a,b}' FROM many_types
 ----
 │
 └ Node 1
-  └ *colexecproj.projJSONFetchValPathJSONDatumConstOp
+  └ *colexecprojconst.projJSONFetchValPathJSONDatumConstOp
     └ *colexecproj.projJSONFetchValPathJSONDatumOp
       └ *colfetcher.ColBatchScan
 
@@ -379,9 +379,9 @@ EXPLAIN (VEC) SELECT _json #> _stringarray #> '{c}', _json #> '{a,b}' #> '{c}' F
 ----
 │
 └ Node 1
-  └ *colexecproj.projJSONFetchValPathJSONDatumConstOp
-    └ *colexecproj.projJSONFetchValPathJSONDatumConstOp
-      └ *colexecproj.projJSONFetchValPathJSONDatumConstOp
+  └ *colexecprojconst.projJSONFetchValPathJSONDatumConstOp
+    └ *colexecprojconst.projJSONFetchValPathJSONDatumConstOp
+      └ *colexecprojconst.projJSONFetchValPathJSONDatumConstOp
         └ *colexecproj.projJSONFetchValPathJSONDatumOp
           └ *colfetcher.ColBatchScan
 
@@ -390,7 +390,7 @@ EXPLAIN (VEC) SELECT _json #>> _stringarray, _json #>> '{a,b}' FROM many_types
 ----
 │
 └ Node 1
-  └ *colexecproj.projJSONFetchTextPathJSONDatumConstOp
+  └ *colexecprojconst.projJSONFetchTextPathJSONDatumConstOp
     └ *colexecproj.projJSONFetchTextPathJSONDatumOp
       └ *colfetcher.ColBatchScan
 
@@ -399,8 +399,8 @@ EXPLAIN (VEC) SELECT _json #> '{a,b}' #>> '{c}' FROM many_types
 ----
 │
 └ Node 1
-  └ *colexecproj.projJSONFetchTextPathJSONDatumConstOp
-    └ *colexecproj.projJSONFetchValPathJSONDatumConstOp
+  └ *colexecprojconst.projJSONFetchTextPathJSONDatumConstOp
+    └ *colexecprojconst.projJSONFetchValPathJSONDatumConstOp
       └ *colfetcher.ColBatchScan
 
 # Make sure we fall back to row engine when we have a mixed-type expression
@@ -420,7 +420,7 @@ EXPLAIN (VEC) SELECT '[2, "hi", {"b": ["bar", {"c": 4}]}]'::jsonb -> _int FROM m
 ----
 │
 └ Node 1
-  └ *colexecproj.projJSONFetchValJSONConstInt64Op
+  └ *colexecprojconst.projJSONFetchValJSONConstInt64Op
     └ *colfetcher.ColBatchScan
 
 # Check that the comparison expressions with the constant on the left are
@@ -431,7 +431,7 @@ EXPLAIN (VEC) SELECT B'11' <> _varbit FROM many_types
 ----
 │
 └ Node 1
-  └ *colexecproj.projNEDatumDatumConstOp
+  └ *colexecprojconst.projNEDatumDatumConstOp
     └ *colfetcher.ColBatchScan
 
 query T
